lil refactoring
